+++ lib/Braintree/Configuration.php
@@ -51,6 +51,12 @@ class Configuration
             if ($kind == 'privateKey') {
                 $this->_privateKey = $value;
             }
+            if ($kind == 'timeout') {
+                $this->_timeout = $value;
+            }
+            if ($kind == 'acceptGzipEncoding') {
+                $this->_acceptGzipEncoding = $value;
+            }
         }
 
         if (isset($attribs['clientId']) || isset($attribs['accessToken'])) {
@@ -414,7 +420,7 @@ class Configuration
         return $this->_sslVersion;
     }
 
-    private function getAcceptGzipEncoding()
+    public function getAcceptGzipEncoding()
     {
         return $this->_acceptGzipEncoding;
     }